offers-sdk/pyproject.toml
[tool.poetry]
name = "offers-sdk"
version = "0.1.0"
description = ""
authors = ["michael-blaha-live <michael.blaha.live@outlook.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = {extras = ["http2"], version = "^0.28.1"}
pydantic = "^2.11.7"
typer = {extras = ["all"], version = "^0.16.0"}
tenacity = "^9.1.2"
pydantic-settings = "^2.10.1"
platformdirs = "^4.3.8"
filelock = "^3.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
respx = "^0.22.0"
python-dotenv = "^1.1.1"
ruff = "^0.12.7"

[tool.poetry.scripts]
offers-cli = "offers_sdk.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

